extract MessageReceiver
